+++ lib/gui_zonegrid.h
@@ -18,7 +18,7 @@ enum roles {
 
 // don't forget whenever we add an item to increase size in NUM_ZONE_ATTRS 
 struct zone {
-  double wave_length; // don't count as ZONE_ATTR, not used in ui
+  int wave_length; // don't count as ZONE_ATTR, not used in ui
   QString name;
   double amp;
   QString origin;
